requirements.txt
numpy<2.0.0
pandas
gspread>=5.10.0
requests
openai
pydub
XlsxWriter
python-dotenv
google-auth-oauthlib
google-api-python-client
google-auth
elevenlabs
chardet
google-cloud-storage
google-cloud-texttospeech
anthropic>=0.7.0
google-generativeai>=0.3.0 